todo_app/pyproject.toml
[project]
name = "todo-app"
version = "1.0.0"
description = "A simple To-Do App built using FastAPI"
packages = [{ include = "app" }]  # Tells Poetry to include the 'app' directory as your package
requires-python = "==3.12.3"
dependencies = [
    "uvicorn>=0.34.3",
    "fastapi>= 0.115.12",
    "sqlalchemy>= 2.0.41",
    "pydantic>= 2.11.5",
    "python-dotenv>= 1.1.0",
    "python-dateutil>= 2.8.2",
    "mcp[cli] (>=1.9.3,<2.0.0)",
    "google-adk>=1.2.1"
]


[tool.poetry]
name = "todo-app"
version = "1.0.0"
description = "A simple To-Do App built using FastAPI"
packages = [{ include = "app" }]  # Tells Poetry to include the 'app' directory as your package

[tool.poetry.dependencies]
python = "3.12.3"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
sqlalchemy = "^2.0.41"
pydantic = "^2.11.5"
python-dotenv = "^1.1.0"
python-dateutil = "^2.8.2"
google-adk = "1.2.1"
mcp = { version = ">=1.9.3,<2.0.0", extras = ["cli"] }



[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"

[tool.poetry.scripts]
local = "app.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
